improved gp (now conditional); moved many functionality to core
