+++ catalyst/contrib/criterion/__init__.py
@@ -11,6 +11,7 @@ from .contrastive import (
 )
 from .dice import BCEDiceLoss, DiceLoss
 from .focal import FocalLossBinary, FocalLossMultiClass
+from .gan import MeanOutputLoss, GradientPenaltyLoss
 from .huber import HuberLoss
 from .iou import BCEIoULoss, IoULoss
 from .lovasz import (
