+++ catalyst/dl/callbacks/criterion.py
@@ -67,7 +67,9 @@ class CriterionCallback(Callback):
         self._get_output = utils.get_dictkey_auto_fn(self.output_key)
         kv_types = (dict, tuple, list, type(None))
         # @TODO: fix to only KV usage
-        if isinstance(self.input_key, str) \
+        if hasattr(self, "_compute_loss"):
+            pass  # overridden in descendants
+        elif isinstance(self.input_key, str) \
                 and isinstance(self.output_key, str):
             self._compute_loss = self._compute_loss_value
         elif isinstance(self.input_key, kv_types) \
@@ -76,25 +78,18 @@ class CriterionCallback(Callback):
         else:
             raise NotImplementedError()
 
-    def _get_additional_criterion_args(self, state: RunnerState):
-        return {}
-
     def _compute_loss_value(self, state: RunnerState, criterion):
         output = self._get_output(state.output, self.output_key)
         input = self._get_input(state.input, self.input_key)
 
-        criterion_kwargs = self._get_additional_criterion_args(state)
-
-        loss = criterion(output, input, **criterion_kwargs)
+        loss = criterion(output, input)
         return loss
 
     def _compute_loss_key_value(self, state: RunnerState, criterion):
         output = self._get_output(state.output, self.output_key)
         input = self._get_input(state.input, self.input_key)
 
-        criterion_kwargs = self._get_additional_criterion_args(state)
-
-        loss = criterion(**output, **input, **criterion_kwargs)
+        loss = criterion(**output, **input)
         return loss
 
     def on_stage_start(self, state: RunnerState):
