+++ examples/mnist_gans/callbacks.py
@@ -1,198 +1,13 @@
 """
 All custom callbacks
 """
-from typing import Any, Callable, Dict, List, Optional, Union
 
 import torch
 import torchvision.utils
 
 from catalyst.dl import RunnerState
-from catalyst.dl.callbacks import CriterionCallback, OptimizerCallback
 from catalyst.dl.core import Callback, CallbackOrder
 from catalyst.utils.tensorboard import SummaryWriter
-
-
-"""
-MetricCallbacks alternatives for input/output keys
-"""
-
-
-class MultiKeyMetricCallback(Callback):
-    """
-    A callback that returns single metric on `state.on_batch_end`
-    """
-
-    # TODO:
-    #  merge it with MetricCallback in catalyst.core
-    #  maybe after the changes with CriterionCallback will be finalized
-    #  in the main repo
-    def __init__(
-        self,
-        prefix: str,
-        metric_fn: Callable,
-        input_key: Optional[Union[str, List[str]]] = "targets",
-        output_key: Optional[Union[str, List[str]]] = "logits",
-        **metric_params
-    ):
-        super().__init__(CallbackOrder.Metric)
-        self.prefix = prefix
-        self.metric_fn = metric_fn
-        self.input_key = input_key
-        self.output_key = output_key
-        self.metric_params = metric_params
-
-    @staticmethod
-    def _get(dictionary: dict, keys: Optional[Union[str, List[str]]]) -> Any:
-        if keys is None:
-            result = dictionary
-        elif isinstance(keys, list):
-            result = {key: dictionary[key] for key in keys}
-        else:
-            result = dictionary[keys]
-        return result
-
-    def on_batch_end(self, state: RunnerState):
-        outputs = self._get(state.output, self.output_key)
-        targets = self._get(state.input, self.input_key)
-        metric = self.metric_fn(outputs, targets, **self.metric_params)
-        state.metrics.add_batch_value(name=self.prefix, value=metric)
-
-
-class WassersteinDistanceCallback(MultiKeyMetricCallback):
-    def __init__(
-        self,
-        prefix: str = "wasserstein_distance",
-        real_validity_output_key: str = "real_validity",
-        fake_validity_output_key: str = "fake_validity"
-    ):
-        super().__init__(
-            prefix,
-            metric_fn=self.get_wasserstein_distance,
-            input_key=None,
-            output_key=[real_validity_output_key, fake_validity_output_key]
-        )
-        self.real_validity_key = real_validity_output_key
-        self.fake_validity_key = fake_validity_output_key
-
-    def get_wasserstein_distance(self, outputs, targets):
-        real_validity = outputs[self.real_validity_key]
-        fake_validity = outputs[self.fake_validity_key]
-        return real_validity.mean() - fake_validity.mean()
-
-
-"""
-CriterionCallback extended
-"""
-
-
-class CriterionWithDiscriminatorCallback(CriterionCallback):
-    """Callback to handle Criterion which has additional argument (model)
-    as input.
-    So imagine you have CRITERION with
-        forward(self, outputs, inputs, discriminator)
-    This callback will add discriminator to criterion forward arguments
-    """
-    def __init__(
-        self,
-        input_key: Union[str, List[str]] = "targets",
-        output_key: Union[str, List[str]] = "logits",
-        prefix: str = "loss",
-        criterion_key: str = None,
-        multiplier: float = 1.0,
-        discriminator_model_key="discriminator",
-        discriminator_model_criterion_key="discriminator"
-    ):
-        """
-
-        :param input_key:
-        :param output_key:
-        :param prefix:
-        :param criterion_key:
-        :param multiplier:
-        :param discriminator_model_key:
-            discriminator key to extract from state.model
-        :param discriminator_model_criterion_key:
-            discriminator key in criterion forward
-            Example 1:
-                forward(self, outputs, inputs, discriminator)
-                (here discriminator_model_criterion_key is "discriminator")
-            Example 2:
-                forward(self, outputs, inputs, d_model)
-                (here discriminator_model_criterion_key is "d_model")
-        """
-        super().__init__(
-            input_key, output_key, prefix, criterion_key, multiplier
-        )
-        self.discriminator_model_key = \
-            discriminator_model_key
-        self.discriminator_model_criterion_key = \
-            discriminator_model_criterion_key
-
-    def _get_additional_criterion_args(self, state: RunnerState):
-        return {
-            self.discriminator_model_criterion_key: state.model[
-                self.discriminator_model_key]
-        }
-
-
-"""
-Optimizer Callback with weights clamp after update
-"""
-
-
-class WeightClampingOptimizerCallback(OptimizerCallback):
-    """
-    Optimizer callback + weights clipping after step is finished
-    """
-    def __init__(
-        self,
-        grad_clip_params: Dict = None,
-        accumulation_steps: int = 1,
-        optimizer_key: str = None,
-        loss_key: str = "loss",
-        decouple_weight_decay: bool = True,
-        weight_clamp_value: float = 0.1
-    ):
-        """
-
-        :param grad_clip_params:
-        :param accumulation_steps:
-        :param optimizer_key:
-        :param loss_key:
-        :param decouple_weight_decay:
-        :param weight_clamp_value:
-            value to clamp weights after each optimization iteration
-            Attention: will clamp WEIGHTS, not GRADIENTS
-        """
-        super().__init__(
-            grad_clip_params=grad_clip_params,
-            accumulation_steps=accumulation_steps,
-            optimizer_key=optimizer_key,
-            loss_key=loss_key,
-            decouple_weight_decay=decouple_weight_decay
-        )
-        self.weight_clamp_value = weight_clamp_value
-
-    def on_batch_end(self, state):
-        """On batch end event"""
-        super().on_batch_end(state)
-        if not state.need_backward:
-            return
-
-        optimizer = state.get_key(
-            key="optimizer", inner_key=self.optimizer_key
-        )
-
-        need_gradient_step = \
-            self._accumulation_counter % self.accumulation_steps == 0
-
-        if need_gradient_step:
-            for group in optimizer.param_groups:
-                for param in group["params"]:
-                    param.data.clamp_(
-                        min=-self.weight_clamp_value,
-                        max=self.weight_clamp_value
-                    )
 
 
 """
@@ -324,7 +139,5 @@ class VisualizationCallback(Callback):
 
 
 __all__ = [
-    "MultiKeyMetricCallback",
-    "WassersteinDistanceCallback", "CriterionWithDiscriminatorCallback",
-    "WeightClampingOptimizerCallback", "VisualizationCallback"
+    "VisualizationCallback"
 ]
